pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faers-longevity-analysis"
version = "0.1.0"
description = "Computational pharmacovigilance analysis of longevity-relevant drugs using FAERS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pharmacovigilance", "FAERS", "drug safety", "longevity", "adverse events"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "pyarrow>=12.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "notebook>=7.0.0",
]
network = [
    "networkx>=3.1",
    "node2vec>=0.4.0",
]
shap = [
    "shap>=0.42.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/faers-longevity-analysis"
Repository = "https://github.com/yourusername/faers-longevity-analysis"
Issues = "https://github.com/yourusername/faers-longevity-analysis/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

